Changed input validation to tool/polyline.py
