+++ src/bonsai/bonsai/bim/module/model/wall.py
@@ -335,7 +335,7 @@ class DrawPolylineWall(bpy.types.Operator):
 
     def recalculate_inputs(self, context):
         if self.number_input:
-            is_valid, self.number_output = tool.Snap.validate_input(self.number_output, self.input_type)
+            is_valid, self.number_output = tool.Polyline.validate_input(self.number_output, self.input_type)
             self.input_ui.set_value(self.input_type, self.number_output)
             if not is_valid:
                 self.report({"WARNING"}, "The number typed is not valid.")
