+++ src/bonsai/bonsai/tool/snap.py
@@ -509,130 +509,3 @@ class Snap(bonsai.core.tool.Snap):
         cls.update_snapping_point(shifted_list[0][0], shifted_list[0][1])
         return shifted_list
 
-    @classmethod
-    def validate_input(cls, input_number, input_type):
-
-        grammar_imperial = """
-        start: (FORMULA dim expr) | dim
-        dim: imperial
-
-        FORMULA: "="
-
-        imperial: feet? "-"? inches?
-        feet: NUMBER? "-"? fraction? "'"
-        inches: NUMBER? "-"? fraction? "\\""
-        fraction: NUMBER "/" NUMBER
-
-        expr: (ADD | SUB) dim | (MUL | DIV) NUMBER
-
-        NUMBER: /-?\\d+(?:\\.\\d+)?/
-        ADD: "+"
-        SUB: "-"
-        MUL: "*"
-        DIV: "/"
-
-        %ignore " "
-        """
-
-        grammar_metric = """
-        start: FORMULA? dim expr?
-        dim: metric
-
-        FORMULA: "="
-
-        metric: NUMBER
-
-        expr: (ADD | SUB | MUL | DIV) dim
-
-        NUMBER: /-?\\d+(?:\\.\\d+)?/
-        ADD: "+"
-        SUB: "-"
-        MUL: "*"
-        DIV: "/"
-
-        %ignore " "
-        """
-
-        class InputTransform(Transformer):
-            def NUMBER(self, n):
-                return float(n)
-
-            def fraction(self, numbers):
-                return numbers[0] / numbers[1]
-
-            def inches(self, args):
-                if len(args) > 1:
-                    result = args[0] + args[1]
-                else:
-                    result = args[0]
-                return result / 12
-
-            def feet(self, args):
-                return args[0]
-
-            def imperial(self, args):
-                if len(args) > 1:
-                    if args[0] <= 0:
-                        result = args[0] - args[1]
-                    else:
-                        result = args[0] + args[1]
-                else:
-                    result = args[0]
-                return result
-
-            def metric(self, args):
-                return args[0]
-
-            def dim(self, args):
-                return args[0]
-
-            def expr(self, args):
-                op = args[0]
-                value = float(args[1])
-                if op == "+":
-                    return lambda x: x + value
-                elif op == "-":
-                    return lambda x: x - value
-                elif op == "*":
-                    return lambda x: x * value
-                elif op == "/":
-                    return lambda x: x / value
-
-            def FORMULA(cls, args):
-                return args[0]
-
-            def start(self, args):
-                i = 0
-                if args[0] == "=":
-                    i += 1
-                else:
-                    if len(args) > 1:
-                        raise ValueError("Invalid input.")
-                dimension = args[i]
-                if len(args) > i + 1:
-                    expression = args[i + 1]
-                    return expression(dimension) * factor
-                else:
-                    return dimension * factor
-
-        try:
-            if bpy.context.scene.unit_settings.system == "IMPERIAL":
-                parser = Lark(grammar_imperial)
-                factor = 0.3048
-            else:
-                parser = Lark(grammar_metric)
-                factor = 1
-                if bpy.context.scene.unit_settings.length_unit == "MILLIMETERS":
-                    factor = 0.001
-
-            if input_type == "A":
-                parser = Lark(grammar_metric)
-                factor = 1
-
-            parse_tree = parser.parse(input_number)
-
-            transformer = InputTransform()
-            result = transformer.transform(parse_tree)
-            return True, str(result)
-        except:
-            return False, "0"
